+++ server/model/src/db/geofence.rs
@@ -281,8 +281,8 @@ impl Query {
     /// Returns paginated Geofence models
     pub async fn paginate(
         db: &DatabaseConnection,
-        page: usize,
-        posts_per_page: usize,
+        page: u64,
+        posts_per_page: u64,
         order: String,
         sort_by: String,
         q: String,
