+++ server/model/src/db/project.rs
@@ -95,8 +95,8 @@ impl Query {
 
     pub async fn paginate(
         db: &DatabaseConnection,
-        page: usize,
-        posts_per_page: usize,
+        page: u64,
+        posts_per_page: u64,
         order: String,
         sort_by: String,
         q: String,
